simplify(payees): extract fuzzyMatch to shared, remove dead code
- Move fuzzyMatch to packages/shared/src/utils/strings.ts; both hub
  page.tsx and mcp-server payees.ts now import from @my-hub/shared/utils
  instead of duplicating the implementation
- Remove redundant `const loading = payeesLoading` alias in page.tsx
- Drop unnecessary wrapper <div> around search Input; apply flex-1 directly
- Sort filtered array in-place (filter already returns a new array)
- Remove `|| submitting` from MergePayeeModal submitDisabled — submitLoading
  already disables the button while submitting
- Delete two WHAT comments in mergePayees service
