+++ packages/shared/CLAUDE.md
@@ -12,6 +12,7 @@
 | `travel.ts`           | Trip domain helpers: `deriveTripStatus` (computes Planned/Active/Completed/Cancelled), `isTransportBookingType`                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       |
 | `finances.ts`         | Finance domain helpers: `getCurrencySymbol(currency)` — maps currency code to display symbol (USD→$, GBP→£, EUR→€), falls back to the code itself; `isPayeeRequired(type)` — returns whether a transaction type should involve payees (false for transfers)                                                                                                                                                                                                                                                                                                                                                                                                           |
 | `logger.ts`           | Minimal logger with ISO timestamps: `logger.info/warn/error` — wraps `console.log/warn/error` with an ISO 8601 prefix. Use in worker and server startup code where no framework logger is available.                                                                                                                                                                                                                                                                                                                                                                                                                                                                  |
+| `strings.ts`          | String utilities: `fuzzyMatch(haystack, needle)` — subsequence fuzzy match; characters must appear in order but need not be adjacent (case-insensitive). Empty needle always returns true.                                                                                                                                                                                                                                                                                                                                                                                                                                                                            |
 
 > Before writing any utility or component, check the tables above. If the function has general purpose (string, number, date, array, etc.), it belongs here even if only one place uses it currently.
