+++ packages/shared/src/utils/index.ts
@@ -4,5 +4,6 @@ export * from './finances';
 export * from './dates';
 export * from './logger';
 export * from './objects';
+export * from './strings';
 export * from './travel';
 export * from './withBackoffRetry';
